refactor(breadcrumbs): move breadcrumbs display logic to breadcrumbs component
