+++ src/app/app.component.html
@@ -1,6 +1,6 @@
 <div class="main-wrapper">
   <app-header></app-header>
-  <app-breadcrumbs [hidden]="!(showBreadcrumbs$ | async)" [steps]="breadcrumbsSteps$ | async"></app-breadcrumbs>
+  <app-breadcrumbs></app-breadcrumbs>
   <main class="content-wrapper">
     <router-outlet></router-outlet>
   </main>
