+++ src/app/app.component.spec.ts
@@ -6,7 +6,6 @@ import { MockBreadcrumbsComponent } from './core/components/breadcrumbs/breadcru
 import { MockFooterComponent } from './core/components/footer/footer.component.mock';
 import { MockHeaderComponent } from './core/components/header/header.component.mock';
 import { SpinnerComponent } from './core/components/spinner/spinner.component';
-import { BreadcrumbsService } from './services/breadcrumbs/breadcrumbs.service';
 
 describe('AppComponent', () => {
   let componentFixture: ComponentFixture<AppComponent>;
@@ -21,7 +20,6 @@ describe('AppComponent', () => {
         MockHeaderComponent,
         SpinnerComponent,
       ],
-      providers: [ BreadcrumbsService ],
     }).compileComponents();
   });
